deps/v8/test/inspector/debugger/stepping-generator-expected.txt

Generator stepping

Running test: testStepOverFromInitialYield
Setting breakpoint on implicit initial yield
Calling callGenerator()

function* generator#() {
  var a = 42;

Stepping over while paused on the initial yield
function callGenerator() {
  return generator();#
}


Running test: testStepIntoInitialYield
Setting breakpoint on call to generator()
Calling callGenerator()
function callGenerator() {
  #return generator();
}

Stepping into the generator()

function* generator#() {
  var a = 42;

Stepping into while paused on the initial yield
function callGenerator() {
  return generator();#
}
